Simplify gemspec and development dependencies
As Rake is only needed to build on Travis CI (and not strictly necessary
for development of the gem), move that dependency to the `Gemfile`.

Remove dependency on Bundler as that can be considered external to the
project (e.g. you might manage gems using only RubyGems).

Use Ruby's standard library `Dir` globbing for finding files in the
gemspec rather than shelling out to `git` which is considerably more
expensive (and requires `git` to be present).

Remove `bin` directory as this is the default path for executable
scripts in gems.